+++ UKSF.Api.Integrations.Teamspeak/Services/TeamspeakMetricsService.cs
@@ -4,12 +4,12 @@ namespace UKSF.Api.Teamspeak.Services
 {
     public interface ITeamspeakMetricsService
     {
-        float GetWeeklyParticipationTrend(HashSet<double> teamspeakIdentities);
+        float GetWeeklyParticipationTrend(HashSet<int> teamspeakIdentities);
     }
 
     public class TeamspeakMetricsService : ITeamspeakMetricsService
     {
-        public float GetWeeklyParticipationTrend(HashSet<double> teamspeakIdentities)
+        public float GetWeeklyParticipationTrend(HashSet<int> teamspeakIdentities)
         {
             return 3;
         }
